feat: Move log types to separe file
