+++ app/logs/page.tsx
@@ -1,15 +1,4 @@
-import LogBox from "@/components/LogBox";
-
-export type Log = {
-  id: string;
-  date: string;
-  foodInput: string[];
-  alcohol: boolean;
-  bowelMovements: string;
-  stress: number;
-  pain: boolean;
-  nausea: boolean;
-};
+import LogBox from '@/components/LogBox';
 
 const Logs = async () => {
   return (
