+++ components/LogBox.tsx
@@ -1,15 +1,14 @@
-"use client";
-import { createClient } from "@supabase/supabase-js";
-import type { Log as LogType } from "../app/logs/page";
-import Link from "next/link";
-import { PlusCircle } from "lucide-react";
+'use client';
+import { createClient } from '@supabase/supabase-js';
+import type { Log as LogType } from '../content/types';
+import Link from 'next/link';
+import { PlusCircle } from 'lucide-react';
+import Log from '@/components/Log';
+import { useEffect, useState } from 'react';
 
 const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL as string;
 const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY as string;
 const supabase = createClient(supabaseUrl, supabaseAnonKey);
-
-import Log from "@/components/Log";
-import { useEffect, useState } from "react";
 
 const LogBox = () => {
   const [logs, setLogs] = useState<LogType[]>([]);
@@ -17,21 +16,21 @@ const LogBox = () => {
   useEffect(() => {
     const getLogs = async () => {
       const { data, error } = await supabase
-        .from("Logs")
+        .from('Logs')
         .select()
-        .order("date", { ascending: false });
+        .order('date', { ascending: false });
       setLogs(data as LogType[]);
     };
     getLogs();
   }, []);
 
-  console.log("logs", logs);
+  console.log('logs', logs);
 
   return (
     <div>
       <div className="flex justify-between m-10">
         <h1 className="text-3xl">Logs</h1>
-        <Link href={"/new-log"}>
+        <Link href={'/new-log'}>
           <PlusCircle size={34} />
         </Link>
       </div>
