+++ content/types.tsx
@@ -1,0 +1,14 @@
+export type Log = {
+  id: string;
+  date: string;
+  foodInput: string[];
+  alcohol: boolean;
+  bowelMovements: string;
+  stress: number;
+  pain: boolean;
+  nausea: boolean;
+};
+
+export type LogProps = {
+  log: Log;
+};
